backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
base58==2.1.1
beautifulsoup4==4.12.3
bitarray==3.0.0
blinker==1.8.2
browser-cookie3==0.19.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cytoolz==1.0.1
distro==1.9.0
dnspython==2.7.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
gemini-api==0.1.6
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
IPFS-Toolkit==0.5.27
ipfshttpclient==0.4.13
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.28
langchain-openai==0.2.14
langsmith==0.2.6
Levenshtein==0.26.0
loguru==0.7.2
lz4==4.3.3
MarkupSafe==3.0.1
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
numpy==2.1.2
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gemini-api==2.4.12
python-Levenshtein==0.26.0
pytz==2024.2
pyunormalize==16.0.0
pywin32==308
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toolz==1.0.0
tqdm==4.66.5
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
varint==1.0.2
web3==7.6.1
websockets==13.1
Werkzeug==3.0.4
win32-setctime==1.1.0
yarl==1.15.4
